Separate entre point class to a different documentation
