+++ entry_point.rb
@@ -1,0 +1,13 @@
+class EntryPoint
+    puts 'Welcome to School Library App'
+  
+    def menu
+      puts '1 - List all books'
+      puts '2 - List all people'
+      puts '3 - Create a person'
+      puts '4 - Create a book'
+      puts '5 - Create a rental'
+      puts '6 - List all rentals for a given person id'
+      puts '7 - Exit'
+    end
+  end
